数据库改动和接口的编写1

+++ src/main/java/com/jcohy/perfectteaching/repository/DeptRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * ClassName  : com.jcohy.perfectteaching.repository
  * Description  :
  */
-public interface DeptRepository  extends JpaRepository<Dept,Long> {
+public interface DeptRepository  extends JpaRepository<Dept,Integer> {
 }
