+++ src/main/java/com/jcohy/perfectteaching/repository/MajorRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * ClassName  : com.jcohy.perfectteaching.repository
  * Description  :
  */
-public interface MajorRepository  extends JpaRepository<Major,Long> {
+public interface MajorRepository  extends JpaRepository<Major,Integer> {
 }
